+++ contracts/protocol/defiAdapters/curve/CurvePoolAdapter.sol
@@ -581,7 +581,7 @@ contract CurvePoolAdapter is IAdapter, Modifiers {
     ) public view override returns (uint256) {
         address[] memory _underlyingTokens = _getUnderlyingTokens(_liquidityPool);
         uint256 tokenIndex = 0;
-        for (uint256 i = 0; i < uint256(_underlyingTokens.length); i++) {
+        for (uint256 i = 0; i < _underlyingTokens.length; i++) {
             if (_underlyingTokens[i] == _underlyingToken) {
                 tokenIndex = i;
             }
@@ -675,7 +675,7 @@ contract CurvePoolAdapter is IAdapter, Modifiers {
     ) public view override returns (uint256) {
         address[] memory _underlyingTokens = _getUnderlyingTokens(_liquidityPool);
         uint256 tokenIndex = 0;
-        for (uint256 i = 0; i < uint256(_underlyingTokens.length); i++) {
+        for (uint256 i = 0; i < _underlyingTokens.length; i++) {
             if (_underlyingTokens[i] == _underlyingToken) {
                 tokenIndex = i;
             }
@@ -914,7 +914,7 @@ contract CurvePoolAdapter is IAdapter, Modifiers {
         if (_amount > 0) {
             address[] memory _underlyingTokens = _getUnderlyingTokens(_liquidityPool);
             uint256 i = 0;
-            for (uint256 j = 0; j < uint256(_underlyingTokens.length); j++) {
+            for (uint256 j = 0; j < _underlyingTokens.length; j++) {
                 if (_underlyingTokens[j] == _underlyingToken) {
                     i = j;
                 }
